Add missing project plugins, partial solution.

+++ examples/ricart-agrawala/src/main/scala/eu/phisikus/plotka/examples/ricart/agrawala/message/AcceptMessage.scala
@@ -1,3 +1,3 @@
 package eu.phisikus.plotka.examples.ricart.agrawala.message
 
-case class AcceptMessage(timestamp: Int)
+case class AcceptMessage(timestamp: Long)
