+++ examples/ricart-agrawala/src/main/scala/eu/phisikus/plotka/examples/ricart/agrawala/message/RequestMessage.scala
@@ -1,3 +1,3 @@
 package eu.phisikus.plotka.examples.ricart.agrawala.message
 
-case class RequestMessage(timestamp : Int)
+case class RequestMessage(timestamp : Long)
